refactor: hooks -> model, services -> lib로 폴더 이름 변경
