+++ apps/frontend/src/features/canvas/model/calculateHandles.ts
