+++ apps/frontend/src/features/editor/model/useEditorTitle.ts
